feat: Update login page UI and functionality
Refactor LogoIcon component to import MainIcon from assets and use it as the image source. This improves performance by avoiding an external HTTP request for the image.

Refactor CreateAccountPopUp component to use react-router-dom's Link component instead of an anchor tag for the "กลับเข้าสู่ระบบ" link. This ensures proper routing within the application.

Refactor ResetPasswordPopUp component to handle form submission and password reset functionality. The handleSubmit function now checks if the new password matches the confirm password before making an API call to reset the password.

These changes enhance the user experience and improve the overall functionality of the login page.