feat: Add getAvatar method to users service
This commit adds a new method, `getAvatar`, to the `usersService` in the `users.service.ts` file. The `getAvatar` method makes a GET request to retrieve the avatar of a user based on their ID. This enhancement improves the functionality of the users service by allowing the retrieval of user avatars.