feat: Add onChange prop to Dropdown component
This commit adds the `onChange` prop to the Dropdown component in Dropdown.tsx. The `onChange` prop is a callback function that is triggered when the selected value in the dropdown changes. This enhancement improves the functionality of the Dropdown component by allowing it to notify the parent component of the selected value.