feat: Update imports for usePopUp and useAuth hooks
This commit updates the imports for the usePopUp and useAuth hooks in multiple files. The imports have been changed from "../../hooks/pop-up.hook" and "../hooks/auth.hook" to "../../../hooks/pop-up.hook" and "../../../hooks/auth.hook" respectively. This change ensures that the correct hooks are being used and improves the clarity of the code.